+++ src/Model/EnvelopeDelayRule.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * BillingInvoiceItem
+ * EnvelopeDelayRule
  *
  * PHP version 7.4
  *
@@ -34,16 +34,15 @@ use \ArrayAccess;
 use DocuSign\eSign\ObjectSerializer;
 
 /**
- * BillingInvoiceItem Class Doc Comment
+ * EnvelopeDelayRule Class Doc Comment
  *
  * @category    Class
- * @description Contains information about an item on a billing invoice.
  * @package     DocuSign\eSign
  * @author      Swagger Codegen team <apihelp@docusign.com>
  * @license     The DocuSign PHP Client SDK is licensed under the MIT License.
  * @link        https://github.com/swagger-api/swagger-codegen
  */
-class BillingInvoiceItem implements ModelInterface, ArrayAccess
+class EnvelopeDelayRule implements ModelInterface, ArrayAccess
 {
     const DISCRIMINATOR = null;
 
@@ -52,7 +51,7 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
       *
       * @var string
       */
-    protected static $swaggerModelName = 'billingInvoiceItem';
+    protected static $swaggerModelName = 'envelopeDelayRule';
 
     /**
       * Array of property to type mappings. Used for (de)serialization
@@ -60,13 +59,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerTypes = [
-        'charge_amount' => '?string',
-        'charge_name' => '?string',
-        'invoice_item_id' => '?string',
-        'quantity' => '?string',
-        'tax_amount' => '?string',
-        'tax_exempt_amount' => '?string',
-        'unit_price' => '?string'
+        'delay' => '?string',
+        'resume_date' => '?string'
     ];
 
     /**
@@ -75,13 +69,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'charge_amount' => null,
-        'charge_name' => null,
-        'invoice_item_id' => null,
-        'quantity' => null,
-        'tax_amount' => null,
-        'tax_exempt_amount' => null,
-        'unit_price' => null
+        'delay' => null,
+        'resume_date' => null
     ];
 
     /**
@@ -111,13 +100,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $attributeMap = [
-        'charge_amount' => 'chargeAmount',
-        'charge_name' => 'chargeName',
-        'invoice_item_id' => 'invoiceItemId',
-        'quantity' => 'quantity',
-        'tax_amount' => 'taxAmount',
-        'tax_exempt_amount' => 'taxExemptAmount',
-        'unit_price' => 'unitPrice'
+        'delay' => 'delay',
+        'resume_date' => 'resumeDate'
     ];
 
     /**
@@ -126,13 +110,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $setters = [
-        'charge_amount' => 'setChargeAmount',
-        'charge_name' => 'setChargeName',
-        'invoice_item_id' => 'setInvoiceItemId',
-        'quantity' => 'setQuantity',
-        'tax_amount' => 'setTaxAmount',
-        'tax_exempt_amount' => 'setTaxExemptAmount',
-        'unit_price' => 'setUnitPrice'
+        'delay' => 'setDelay',
+        'resume_date' => 'setResumeDate'
     ];
 
     /**
@@ -141,13 +120,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $getters = [
-        'charge_amount' => 'getChargeAmount',
-        'charge_name' => 'getChargeName',
-        'invoice_item_id' => 'getInvoiceItemId',
-        'quantity' => 'getQuantity',
-        'tax_amount' => 'getTaxAmount',
-        'tax_exempt_amount' => 'getTaxExemptAmount',
-        'unit_price' => 'getUnitPrice'
+        'delay' => 'getDelay',
+        'resume_date' => 'getResumeDate'
     ];
 
     /**
@@ -210,13 +184,8 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['charge_amount'] = isset($data['charge_amount']) ? $data['charge_amount'] : null;
-        $this->container['charge_name'] = isset($data['charge_name']) ? $data['charge_name'] : null;
-        $this->container['invoice_item_id'] = isset($data['invoice_item_id']) ? $data['invoice_item_id'] : null;
-        $this->container['quantity'] = isset($data['quantity']) ? $data['quantity'] : null;
-        $this->container['tax_amount'] = isset($data['tax_amount']) ? $data['tax_amount'] : null;
-        $this->container['tax_exempt_amount'] = isset($data['tax_exempt_amount']) ? $data['tax_exempt_amount'] : null;
-        $this->container['unit_price'] = isset($data['unit_price']) ? $data['unit_price'] : null;
+        $this->container['delay'] = isset($data['delay']) ? $data['delay'] : null;
+        $this->container['resume_date'] = isset($data['resume_date']) ? $data['resume_date'] : null;
     }
 
     /**
@@ -244,169 +213,49 @@ class BillingInvoiceItem implements ModelInterface, ArrayAccess
 
 
     /**
-     * Gets charge_amount
+     * Gets delay
      *
      * @return ?string
      */
-    public function getChargeAmount()
+    public function getDelay()
     {
-        return $this->container['charge_amount'];
+        return $this->container['delay'];
     }
 
     /**
-     * Sets charge_amount
+     * Sets delay
      *
-     * @param ?string $charge_amount Reserved: TBD
+     * @param ?string $delay 
      *
      * @return $this
      */
-    public function setChargeAmount($charge_amount)
+    public function setDelay($delay)
     {
-        $this->container['charge_amount'] = $charge_amount;
+        $this->container['delay'] = $delay;
 
         return $this;
     }
 
     /**
-     * Gets charge_name
+     * Gets resume_date
      *
      * @return ?string
      */
-    public function getChargeName()
+    public function getResumeDate()
     {
-        return $this->container['charge_name'];
+        return $this->container['resume_date'];
     }
 
     /**
-     * Sets charge_name
+     * Sets resume_date
      *
-     * @param ?string $charge_name Reserved: TBD
+     * @param ?string $resume_date 
      *
      * @return $this
      */
-    public function setChargeName($charge_name)
+    public function setResumeDate($resume_date)
     {
-        $this->container['charge_name'] = $charge_name;
-
-        return $this;
-    }
-
-    /**
-     * Gets invoice_item_id
-     *
-     * @return ?string
-     */
-    public function getInvoiceItemId()
-    {
-        return $this->container['invoice_item_id'];
-    }
-
-    /**
-     * Sets invoice_item_id
-     *
-     * @param ?string $invoice_item_id Reserved: TBD
-     *
-     * @return $this
-     */
-    public function setInvoiceItemId($invoice_item_id)
-    {
-        $this->container['invoice_item_id'] = $invoice_item_id;
-
-        return $this;
-    }
-
-    /**
-     * Gets quantity
-     *
-     * @return ?string
-     */
-    public function getQuantity()
-    {
-        return $this->container['quantity'];
-    }
-
-    /**
-     * Sets quantity
-     *
-     * @param ?string $quantity 
-     *
-     * @return $this
-     */
-    public function setQuantity($quantity)
-    {
-        $this->container['quantity'] = $quantity;
-
-        return $this;
-    }
-
-    /**
-     * Gets tax_amount
-     *
-     * @return ?string
-     */
-    public function getTaxAmount()
-    {
-        return $this->container['tax_amount'];
-    }
-
-    /**
-     * Sets tax_amount
-     *
-     * @param ?string $tax_amount 
-     *
-     * @return $this
-     */
-    public function setTaxAmount($tax_amount)
-    {
-        $this->container['tax_amount'] = $tax_amount;
-
-        return $this;
-    }
-
-    /**
-     * Gets tax_exempt_amount
-     *
-     * @return ?string
-     */
-    public function getTaxExemptAmount()
-    {
-        return $this->container['tax_exempt_amount'];
-    }
-
-    /**
-     * Sets tax_exempt_amount
-     *
-     * @param ?string $tax_exempt_amount 
-     *
-     * @return $this
-     */
-    public function setTaxExemptAmount($tax_exempt_amount)
-    {
-        $this->container['tax_exempt_amount'] = $tax_exempt_amount;
-
-        return $this;
-    }
-
-    /**
-     * Gets unit_price
-     *
-     * @return ?string
-     */
-    public function getUnitPrice()
-    {
-        return $this->container['unit_price'];
-    }
-
-    /**
-     * Sets unit_price
-     *
-     * @param ?string $unit_price Reserved: TBD
-     *
-     * @return $this
-     */
-    public function setUnitPrice($unit_price)
-    {
-        $this->container['unit_price'] = $unit_price;
+        $this->container['resume_date'] = $resume_date;
 
         return $this;
     }